*修改了64位下和Release版编译的问题 *log info强制使用灰色文本输出。

+++ clstd/clFile.cpp
@@ -315,7 +315,7 @@ namespace clstd
 
   b32 File::Write(const BufferBase& buffer, u32* lpNumberOfBytesWritten /*= NULL*/)
   {
-    return Write(buffer.GetPtr(), buffer.GetSize(), lpNumberOfBytesWritten);
+    return Write(buffer.GetPtr(), (u32)buffer.GetSize(), lpNumberOfBytesWritten);
   }
 
   //int File::Writef(const tch* format, ...)
